move p4rt test back to test.py
